requirements.txt
Flask>=2.0
flask-cors>=3.0
flask-jwt-extended>=4.0
mysql-connector-python>=8.0
flask-bcrypt>=1.0
gunicorn>=20.0
